requirements.txt
discord.py==2.2.2
nest_asyncio==1.5.5
openai==0.27.2
